refactor(settings): simplifies/cleans up
