+++ src/v2/Apps/Settings/Routes/EditSettings/Components/__tests__/SettingsEditSettingsInformation.jest.tsx
@@ -1,7 +1,7 @@
 import { screen, fireEvent } from "@testing-library/react"
 import { graphql } from "react-relay"
 import { setupTestWrapperTL } from "v2/DevTools/setupTestWrapper"
-import { SettingsEditSettingsInformationFragmentContainer } from "../Components/SettingsEditSettingsInformation"
+import { SettingsEditSettingsInformationFragmentContainer } from "../SettingsEditSettingsInformation"
 
 jest.unmock("react-relay")
