+++ src/v2/Apps/Settings/Routes/EditSettings/SettingsEditSettingsRoute.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import { SettingsEditSettingsRoute_me } from "v2/__generated__/SettingsEditSettingsRoute_me.graphql"
 import { createFragmentContainer, graphql } from "react-relay"
-import { SettingsEditSettingsInformationFragmentContainer } from "../EditSettings/Components/SettingsEditSettingsInformation"
-import { SettingsEditSettingsTwoFactorRefetchContainer } from "./Components/SettingsEditSettingsTwoFactor/SettingsEditSettingsTwoFactor"
+import { SettingsEditSettingsInformationFragmentContainer } from "./Components/SettingsEditSettingsInformation"
+import { SettingsEditSettingsTwoFactorFragmentContainer } from "./Components/SettingsEditSettingsTwoFactor"
 import { Column, GridColumns, Join, Separator } from "@artsy/palette"
 
 interface SettingsEditRouteProps {
@@ -18,7 +18,7 @@ const SettingsEditRoute: React.FC<SettingsEditRouteProps> = ({ me }) => {
 
           {/* TODO: SettingsEditSettingsPassword */}
 
-          <SettingsEditSettingsTwoFactorRefetchContainer me={me}/>
+          <SettingsEditSettingsTwoFactorFragmentContainer me={me} />
 
           {/* TODO: SettingsEditSettingsLinkedAccounts */}
